requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
scipy>=1.7.0

# Azure ML SDK
azureml-core>=1.45.0
azureml-train-core>=1.45.0
azureml-pipeline>=1.45.0
azureml-pipeline-steps>=1.45.0
azureml-dataset-runtime>=1.45.0

# MLflow for Experiment Tracking
mlflow>=2.0.0

# Data Validation and Monitoring
great-expectations>=0.15.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
joblib>=1.1.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code Quality
black>=22.0.0
flake8>=4.0.0
pylint>=2.12.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Additional ML Tools
shap>=0.40.0  # For model explainability
imbalanced-learn>=0.9.0  # For handling imbalanced datasets
